+++ src/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: dcelojev <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/29 11:04:59 by dcelojev          #+#    #+#             */
-/*   Updated: 2019/04/29 11:09:27 by dcelojev         ###   ########.fr       */
+/*   Updated: 2019/05/03 12:23:07 by dcelojev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = -1;
-	while (++i < (int)n)
+	while (++i < n)
 		((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
 	return (dst);
 }
