+++ src/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: dcelojev <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/29 18:33:55 by dcelojev          #+#    #+#             */
-/*   Updated: 2019/04/30 21:53:48 by dcelojev         ###   ########.fr       */
+/*   Updated: 2019/05/03 12:27:19 by dcelojev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	int i;
+	size_t i;
 
 	i = -1;
-	while (src[++i] && i < (int)len)
+	while (src[++i] && i < len)
 		dst[i] = src[i];
-	while (i < (int)len)
+	while (i < len)
 		dst[i++] = '\0';
 	return (dst);
 }
